Vue component: MainModeSelector
